requirements.txt
accelerate==1.0.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
attrs==25.1.0
av==12.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.1.1
cycler==0.12.1
datasets==3.1.0
deepspeed==0.15.0
dill==0.3.8
docstring_parser==0.16
einops==0.8.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.11
ffmpy==0.5.0
filelock==3.16.1
fire==0.7.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.9.0
gradio==4.44.1
gradio_client==1.3.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
importlib_resources==6.4.5
jieba==0.42.1
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
# Editable install with no version control (llamafactory==0.9.1)
-e YOUR_ROOT_DIR/Hidden_Ghost_Hand/LLaMA-Factory
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.1
ninja==1.11.1.3
nltk==3.9.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
orjson==3.10.15
packaging==24.2
pandas==2.0.3
peft==0.12.0
pillow==10.4.0
propcache==0.2.0
protobuf==5.29.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pynvml==11.5.3
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
qwen-vl-utils==0.0.10
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rouge-chinese==1.0.3
ruff==0.9.10
safetensors==0.5.3
scipy==1.10.1
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
shtab==1.7.1
six==1.17.0
sniffio==1.3.1
sse-starlette==2.1.3
starlette==0.44.0
sympy==1.13.3
termcolor==2.4.0
tiktoken==0.7.0
tokenizers==0.20.3
tomlkit==0.12.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
transformers==4.46.1
triton==3.0.0
trl==0.9.6
typer==0.15.2
typing_extensions==4.12.2
tyro==0.8.14
tzdata==2025.1
urllib3==2.2.3
uvicorn==0.33.0
websockets==12.0
xxhash==3.5.0
yarl==1.15.2
zipp==3.20.2
